+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Switch`: handle `aria-disabled=true` for an accessible disabled state
 -   `DisabledStateProvider` and `useDisabledStateContext` utilities to set disabled state in nested children
 -   `Thumbnail`: handle disabled state context
+-   `Uploader`: handle disabled state (via context, `isDisabled` and `aria-disabled`)
 
 ### Fixed